backend/foodgram/requirements.txt
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.2.2
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
distro==1.7.0
Django==3.2.14
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-extra-fields==3.4.0
flake8==4.0.1
idna==3.3
importlib-metadata==1.7.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
paramiko==2.11.0
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
pywin32==227
PyYAML==5.4.1
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
texttable==1.6.4
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
websocket-client==0.59.0
zipp==3.8.0
